refactor the Transpyler.translate function
