+++ src/transpyler/translate/__init__.py
@@ -1,3 +1,3 @@
 from .gettext import gettext_for, create_pot_file, L10N_PATH
 from .extract import extract_translations, extract_translation
-from .translate import translate_namespace, apply_translations
+from .translate import translate_namespace, apply_translations, translator_factory
